+++ HISTORY.md
@@ -1,14 +1,16 @@
 Work in progress
 ----------------
-- new packages: glossaries, glossaries-extra
+- new LaTeX packages: glossaries, glossaries-extra
+- LaTeX builtins: added macro \\index
 - list-of-macros.md: updated
 - yalafi core
     - allow '@' in macro names (parameters.py:Parameters.macro\_characters())
     - added method parser.py:Parser.parse\_keyval\_list(): parses lists like
       `name=xxx, description={a b c}`
     - non-alphanumeric characters (except '.') in LaTeX package names are
-      replaced by '\_' for corresponding Python package names
+      replaced by '\_' for corresponding Python module names
       (utils.py:get\_module\_handler())
+- travis.yml: use 'python -m pytest' instead of 'pytest'
 
 Version 1.1.3 (2020/08/07)
 --------------------------