+++ HISTORY.md
@@ -11,6 +11,7 @@ Work in progress
       replaced by '\_' for corresponding Python module names
       (utils.py:get\_module\_handler())
 - travis.yml: use 'python -m pytest' instead of 'pytest'
+- added copy of vimtex compiler script vlty.vim
 
 Version 1.1.3 (2020/08/07)
 --------------------------